tests/testing-requirements.txt
pytest-aiohttp==0.3.0
pytest-asyncio==0.10.0
asynctest==0.13.0
aioresponses==0.6.1
pytest-cov==2.8.1
requests-mock==1.7.0